refactor : nettoyage structurel du repo (audit, dette technique)
Suite à l'audit `docs/audit/complexite-2026-04.md`, application des leviers
à plus fort ROI / plus faible risque. Aucun impact calculateur (1571 tests,
npm run check OK).

Nettoyage legacy
- Supprime `src/features/per/styles/legacy.css` (fusion dans `steps.css`).
- Renomme `successionDraft.legacy.ts` → `successionDraft.migrations.ts`
  (nom qui reflète vraiment la fonction : migrations v1→v27 actives).

Découpage fichier monstre
- Éclate `src/pptx/designSystem/serenity.ts` (1 032 l.) en 8 modules ciblés
  (layout, typography, coords, logo, chapters, colorsShadows, helpers,
  components). `serenity.ts` devient un barrel de 19 lignes : tous les
  consommateurs (25) restent inchangés.

Centralisation fixtures tests succession
- Nouveau `src/features/succession/__tests__/fixtures.ts` : `makeCivil`,
  `makeCivilMarie`, `makeDevolution`, `makeLiquidation`.
- 13 fichiers de tests qui dupliquaient ces makers inline pointent
  désormais dessus. Net : -167 lignes.

Homogénéisation structure modulaire (cahier des charges unique)
Alignement sur le modèle credit/placement (main `.tsx` à la racine,
sous-dossiers typés) :
- succession : `successionXlsx.ts` → `export/successionXlsx.ts`
- audit : `exportAudit.ts` → `export/`, `storage.ts` → `utils/`,
  `steps/` → `components/steps/`
- strategy : `exportStrategy.ts` → `export/`, `calculations.ts` +
  `recommendations.ts` → `utils/`, `StrategyBuilder.css` → `styles/`

Livrable audit
- `docs/audit/complexite-2026-04.md` : diagnostic reproductible
  (85 756 l. / 483 fichiers, top 3 leviers ROI, risque borné par cible).
