+++ src/features/per/styles/page.css
@@ -2,5 +2,4 @@
 @import './wizard.css';
 @import './steps.css';
 @import './summary.css';
-@import './legacy.css';
 @import './responsive.css';
