+++ src/features/strategy/index.ts
@@ -3,7 +3,7 @@
  */
 
 export { default as StrategyBuilder } from './StrategyBuilder';
-export * from './exportStrategy';
+export * from './export/exportStrategy';
 export * from './types';
-export * from './recommendations';
-export * from './calculations';
+export * from './utils/recommendations';
+export * from './utils/calculations';
