+++ src/features/strategy/styles/StrategyBuilder.css
@@ -1,5 +1,5 @@
-@import '../../styles/sim/fields.css';
-@import '../../styles/sim/modals.css';
+@import '../../../styles/sim/fields.css';
+@import '../../../styles/sim/modals.css';
 
 /* StrategyBuilder Styles */
