+++ src/features/succession/__tests__/successionDisplay.test.ts
@@ -1,18 +1,12 @@
 import { describe, expect, it } from 'vitest';
-import { DEFAULT_DMTG, type RegimeMatrimonial } from '../../../engine/civil';
+import { DEFAULT_DMTG } from '../../../engine/civil';
 import { calculateSuccession } from '../../../engine/succession';
-import type {
-  SuccessionCivilContext,
-  SuccessionDevolutionContext,
-  SuccessionDevolutionContextInput,
-  SuccessionLiquidationContext,
-} from '../successionDraft';
-import { DEFAULT_SUCCESSION_DEVOLUTION_CONTEXT } from '../successionDraft';
 import { buildSuccessionDevolutionAnalysis } from '../successionDevolution';
 import {
   buildSuccessionDirectDisplayAnalysis,
   computeSuccessionDirectEstateBasis,
 } from '../successionDisplay';
+import { makeCivil, makeDevolution, makeLiquidation } from './fixtures';
 
 const DONATION_SETTINGS = {
   rappelFiscalAnnees: 15,
@@ -24,50 +18,6 @@ const DONATION_SETTINGS = {
     abattementRenouvellement: 15,
   },
 } as const;
-
-function makeCivil(overrides: Partial<SuccessionCivilContext>): SuccessionCivilContext {
-  return {
-    situationMatrimoniale: 'celibataire',
-    regimeMatrimonial: null as RegimeMatrimonial | null,
-    pacsConvention: 'separation',
-    ...overrides,
-  };
-}
-
-function makeDevolution(overrides: SuccessionDevolutionContextInput): SuccessionDevolutionContext {
-  return {
-    ...DEFAULT_SUCCESSION_DEVOLUTION_CONTEXT,
-    ...overrides,
-    testamentsBySide: {
-      ...DEFAULT_SUCCESSION_DEVOLUTION_CONTEXT.testamentsBySide,
-      ...overrides.testamentsBySide,
-      epoux1: {
-        ...DEFAULT_SUCCESSION_DEVOLUTION_CONTEXT.testamentsBySide.epoux1,
-        ...overrides.testamentsBySide?.epoux1,
-        particularLegacies: overrides.testamentsBySide?.epoux1?.particularLegacies ?? [],
-      },
-      epoux2: {
-        ...DEFAULT_SUCCESSION_DEVOLUTION_CONTEXT.testamentsBySide.epoux2,
-        ...overrides.testamentsBySide?.epoux2,
-        particularLegacies: overrides.testamentsBySide?.epoux2?.particularLegacies ?? [],
-      },
-    },
-    ascendantsSurvivantsBySide: {
-      ...DEFAULT_SUCCESSION_DEVOLUTION_CONTEXT.ascendantsSurvivantsBySide,
-      ...overrides.ascendantsSurvivantsBySide,
-    },
-  };
-}
-
-function makeLiquidation(overrides: Partial<SuccessionLiquidationContext>): SuccessionLiquidationContext {
-  return {
-    actifEpoux1: 0,
-    actifEpoux2: 0,
-    actifCommun: 0,
-    nbEnfants: 0,
-    ...overrides,
-  };
-}
 
 describe('buildSuccessionDirectDisplayAnalysis', () => {
   it('calcule des droits directs non nuls pour un celibataire avec deux enfants', () => {
