+++ src/features/succession/__tests__/successionValidationMatrix.test.ts
@@ -1,13 +1,10 @@
 import { describe, expect, it } from 'vitest';
-import { DEFAULT_DMTG, type RegimeMatrimonial } from '../../../engine/civil';
+import { DEFAULT_DMTG } from '../../../engine/civil';
 import type {
   SuccessionCivilContext,
-  SuccessionDevolutionContext,
-  SuccessionDevolutionContextInput,
   SuccessionEnfant,
   SuccessionLiquidationContext,
 } from '../successionDraft';
-import { DEFAULT_SUCCESSION_DEVOLUTION_CONTEXT } from '../successionDraft';
 import { buildSuccessionChainageAnalysis } from '../successionChainage';
 import { buildSuccessionDevolutionAnalysis } from '../successionDevolution';
 import {
@@ -16,50 +13,7 @@ import {
   computeSuccessionDirectEstateBasis,
 } from '../successionDisplay';
 import type { SuccessionAssetTransmissionBasis } from '../successionTransmissionBasis';
-
-function makeCivil(overrides: Partial<SuccessionCivilContext>): SuccessionCivilContext {
-  return {
-    situationMatrimoniale: 'celibataire',
-    regimeMatrimonial: null as RegimeMatrimonial | null,
-    pacsConvention: 'separation',
-    ...overrides,
-  };
-}
-
-function makeDevolution(overrides: SuccessionDevolutionContextInput): SuccessionDevolutionContext {
-  return {
-    ...DEFAULT_SUCCESSION_DEVOLUTION_CONTEXT,
-    ...overrides,
-    testamentsBySide: {
-      ...DEFAULT_SUCCESSION_DEVOLUTION_CONTEXT.testamentsBySide,
-      ...overrides.testamentsBySide,
-      epoux1: {
-        ...DEFAULT_SUCCESSION_DEVOLUTION_CONTEXT.testamentsBySide.epoux1,
-        ...overrides.testamentsBySide?.epoux1,
-        particularLegacies: overrides.testamentsBySide?.epoux1?.particularLegacies ?? [],
-      },
-      epoux2: {
-        ...DEFAULT_SUCCESSION_DEVOLUTION_CONTEXT.testamentsBySide.epoux2,
-        ...overrides.testamentsBySide?.epoux2,
-        particularLegacies: overrides.testamentsBySide?.epoux2?.particularLegacies ?? [],
-      },
-    },
-    ascendantsSurvivantsBySide: {
-      ...DEFAULT_SUCCESSION_DEVOLUTION_CONTEXT.ascendantsSurvivantsBySide,
-      ...overrides.ascendantsSurvivantsBySide,
-    },
-  };
-}
-
-function makeLiquidation(overrides: Partial<SuccessionLiquidationContext>): SuccessionLiquidationContext {
-  return {
-    actifEpoux1: 0,
-    actifEpoux2: 0,
-    actifCommun: 0,
-    nbEnfants: 0,
-    ...overrides,
-  };
-}
+import { makeCivil, makeDevolution, makeLiquidation } from './fixtures';
 
 const DEUX_ENFANTS_DEFUNT: SuccessionEnfant[] = [
   { id: 'E1', rattachement: 'epoux1' },
