+++ src/pptx/designSystem/serenity.ts
@@ -1,1032 +1,19 @@
 /**
  * Serenity Design System for PPTX Generation
- * 
+ *
  * All coordinates in inches for 16:9 widescreen (13.3333" x 7.5")
  * Font: Arial ONLY
  * Colors: From theme roles (no hardcoded except white #FFFFFF)
- */
-
-import type PptxGenJS from 'pptxgenjs';
-import type { PptxThemeRoles, LogoPlacement } from '../theme/types';
-
-// ============================================================================
-// SLIDE SIZE
-// ============================================================================
-
-export const SLIDE_SIZE = {
-  width: 13.3333,
-  height: 7.5,
-  layout: 'LAYOUT_WIDE' as const,
-} as const;
-
-// ============================================================================
-// LOGO PLACEMENT CONSTANTS
-// Safe zone calculations for logo positioning on cover slide
-// ============================================================================
-
-export const LOGO_PLACEMENT = {
-  /** Margin from top edge for 'top' positions */
-  marginTop: 0.5,
-  /** Margin before title zone for 'bottom' positions */
-  marginBottom: 0.3,
-  /** Side margin for left/right positions */
-  marginSide: 0.5,
-  /** Maximum Y for logo bottom (title zone Y - marginBottom) */
-  get maxBottomY() {
-    return COORDS_COVER.title.y - this.marginBottom;
-  },
-  /** Center X of slide */
-  centerX: SLIDE_SIZE.width / 2,
-} as const;
-
-/** Default logo placement */
-export const DEFAULT_LOGO_PLACEMENT: LogoPlacement = 'center-bottom';
-
-/**
- * Calculate logo position based on placement option
- * Ensures logo never overlaps with title zone
- */
-export function calculateLogoPosition(
-  placement: LogoPlacement,
-  logoWidth: number,
-  logoHeight: number
-): { x: number; y: number } {
-  switch (placement) {
-    case 'center-top':
-      return {
-        x: LOGO_PLACEMENT.centerX - logoWidth / 2,
-        y: LOGO_PLACEMENT.marginTop,
-      };
-    case 'top-left':
-      return {
-        x: LOGO_PLACEMENT.marginSide,
-        y: LOGO_PLACEMENT.marginTop,
-      };
-    case 'top-right':
-      return {
-        x: SLIDE_SIZE.width - logoWidth - LOGO_PLACEMENT.marginSide,
-        y: LOGO_PLACEMENT.marginTop,
-      };
-    case 'center-bottom':
-      return {
-        x: LOGO_PLACEMENT.centerX - logoWidth / 2,
-        y: LOGO_PLACEMENT.maxBottomY - logoHeight,
-      };
-    case 'bottom-left':
-      return {
-        x: LOGO_PLACEMENT.marginSide,
-        y: LOGO_PLACEMENT.maxBottomY - logoHeight,
-      };
-    case 'bottom-right':
-      return {
-        x: SLIDE_SIZE.width - logoWidth - LOGO_PLACEMENT.marginSide,
-        y: LOGO_PLACEMENT.maxBottomY - logoHeight,
-      };
-    default:
-      // Fallback to center-bottom
-      return {
-        x: LOGO_PLACEMENT.centerX - logoWidth / 2,
-        y: LOGO_PLACEMENT.maxBottomY - logoHeight,
-      };
-  }
-}
-
-// ============================================================================
-// TYPOGRAPHY
-// ============================================================================
-
-export const TYPO = {
-  fontFace: 'Arial',
-  sizes: {
-    h1: 24,
-    h2: 16,
-    body: 14,
-    bodySmall: 12,
-    bodyXSmall: 10,
-    legal: 11,
-    footer: 8,
-    coverTitle: 24,
-    coverSubtitle: 18,
-    coverMeta: 14, // Changed from 18 to 14 per spec
-  },
-} as const;
-
-// ============================================================================
-// COORDINATES - COVER SLIDE
-// ============================================================================
-
-export const COORDS_COVER = {
-  logo: { x: 4.4844, y: 1.9542, w: 4.3646, h: 1.9896 },
-  title: { x: 1.5528, y: 4.0986, w: 10.2277, h: 0.8333 },
-  dividerLine: { x: 4.4520, y: 5.0778, w: 4.4294 },
-  subtitle: { x: 1.5528, y: 5.2418, w: 10.2277, h: 0.5443 },
-  metaLeft: { x: 0.9784, y: 6.0417, w: 2.8646, h: 0.8333 },
-  metaRight: { x: 9.4903, y: 6.0417, w: 2.8646, h: 0.8333 },
-  cornerMarkLeft: { x: 0.8229, y: 6.1459, w: 0.9063, h: 0.9062 },
-  cornerMarkRight: { x: 11.6042, y: 6.1458, w: 0.9063, h: 0.9062 },
-} as const;
-
-// ============================================================================
-// COORDINATES - CHAPTER SLIDE
-// ============================================================================
-
-export const COORDS_CHAPTER = {
-  panel: { x: 0.5966, y: 0.7347, w: 12.14, h: 5.8704 },
-  image: { x: 0.5966, y: 0.7347, w: 4.2424, h: 5.8704 },
-  title: { x: 4.9909, y: 0.9223, w: 7.3319, h: 0.8663 },
-  subtitle: { x: 4.9909, y: 1.9535, w: 7.3319, h: 0.4149 },
-  accentLine: { x: 5.0818, y: 1.7886, w: 1.1278 },
-  body: { x: 4.9909, y: 2.5334, w: 7.3319, h: 3.7227 },
-} as const;
-
-// ============================================================================
-// COORDINATES - CONTENT SLIDE
-// ============================================================================
-
-export const COORDS_CONTENT = {
-  margin: { x: 0.9167, w: 11.5 },
-  title: { x: 0.9167, y: 0.7643, w: 11.5, h: 0.8663 },
-  subtitle: { x: 0.9167, y: 1.7956, w: 11.5, h: 0.4149 },
-  accentLine: { x: 1.0029, y: 1.6408, w: 1.0614 },
-  content: { x: 0.9167, y: 2.3754, w: 11.5, h: 4.3602 },
-} as const;
-
-// ============================================================================
-// COORDINATES - END SLIDE
-// ============================================================================
-
-export const COORDS_END = {
-  // Legal block centered horizontally
-  legalBlock: { x: 2.6458, y: 2.0166, w: 8.0417, h: 3.4669 },
-  // Corner marks computed from CORNER_MARKS constants for perfect symmetry
-  // topRight: x = slideWidth - marginX - size, y = marginY
-  // bottomLeft: x = marginX, y = slideHeight - marginY - size
-} as const;
-
-// ============================================================================
-// COORDINATES - FOOTER (shared)
-// ============================================================================
-
-export const COORDS_FOOTER = {
-  date: { x: 0.9167, y: 6.9514, w: 1.6875, h: 0.3993 },
-  disclaimer: { x: 2.9792, y: 6.9514, w: 7.375, h: 0.3993 },
-  slideNum: { x: 10.7292, y: 6.9514, w: 1.6875, h: 0.3993 },
-} as const;
-
-// ============================================================================
-// LINE STYLES
-// ============================================================================
-
-export const LINE_STYLES = {
-  divider: { w: 0.75 / 72 }, // 0.75 pt in inches
-  accent: { w: 1.5 / 72 },   // 1.5 pt in inches
-} as const;
-
-// ============================================================================
-// RADIUS TOKENS (shared for panel/image harmony)
-// ============================================================================
-
-export const RADIUS = {
-  /** 
-   * Panel radius (rectRadius for roundRect shape)
-   * Must match the pre-processed image corner radius for seamless junction
-   */
-  panel: 0.12,
-  /** 
-   * Image radius adjustment (same as panel for perfect alignment)
-   * Used when clipping/masking is applied
-   */
-  imageAdj: 0.12,
-} as const;
-
-// ============================================================================
-// BLEED TOKENS (anti-aliasing gap elimination)
-// ============================================================================
-
-export const BLEED = {
-  /**
-   * Image bleed (in inches) - image extends slightly under the panel border
-   * This eliminates the "hole" caused by anti-aliasing at rounded corners
-   * Value 0.02 (~1.5px) is enough to hide gaps without visual distortion
-   */
-  image: 0.02,
-} as const;
-
-// ============================================================================
-// CORNER MARKS LAYOUT (symmetric positioning)
-// Computed mathematically for perfect symmetry
-// ============================================================================
-
-export const CORNER_MARKS = {
-  /** Size of the corner mark group (width = height) */
-  size: 0.65,
-  /** Horizontal margin from slide edge (same for left and right) */
-  marginX: 0.75,
-  /** Vertical margin from slide edge (same for top and bottom) */
-  marginY: 0.75,
-  /** Line spacing between the two vertical lines */
-  lineSpacing: 0.12,
-  /** Height of primary line (inches) */
-  primaryHeight: 0.55,
-  /** Height of secondary line (inches) */
-  secondaryHeight: 0.40,
-  /** Vertical offset for secondary line */
-  secondaryOffset: 0.08,
-  /** Line width in points */
-  lineWidth: 0.75,
-} as const;
-
-// ============================================================================
-// LAYOUT CONTRACT (strict bounding boxes for each zone)
-// Prevents overflow and encroachment on title/subtitle areas
-// ============================================================================
-
-export const LAYOUT_ZONES = {
-  /** Chapter slide zones */
-  chapter: {
-    titleBox: { x: 4.9909, y: 0.9223, w: 7.3319, h: 0.8663 },
-    subtitleBox: { x: 4.9909, y: 1.9535, w: 7.3319, h: 0.6 },
-    bodyBox: { x: 4.9909, y: 2.6, w: 7.3319, h: 3.6 },
-  },
-  /** Content slide zones */
-  content: {
-    titleBox: { x: 0.9167, y: 0.7643, w: 11.5, h: 0.8663 },
-    subtitleBox: { x: 0.9167, y: 1.7956, w: 11.5, h: 0.4149 },
-    bodyBox: { x: 0.9167, y: 2.3754, w: 11.5, h: 4.3602 },
-  },
-  /** Cover slide zones */
-  cover: {
-    titleBox: { x: 1.5528, y: 4.0986, w: 10.2277, h: 0.8333 },
-    subtitleBox: { x: 1.5528, y: 5.2418, w: 10.2277, h: 0.5443 },
-  },
-} as const;
-
-/** Minimum font sizes for text fitting (prevents unreadable text) */
-export const MIN_FONT_SIZES = {
-  h1: 18,
-  h2: 12,
-  body: 10,
-  bodySmall: 8,
-} as const;
-
-// ============================================================================
-// TEXT-BASED LAYOUT TOKENS (for accent line positioning)
-// ============================================================================
-
-/** Gaps for text-based layout (in inches) */
-export const TEXT_GAPS = {
-  /** Gap after title text before accent line */
-  afterTitle: 0.1,
-  /** Gap after accent line before subtitle */
-  afterLine: 0.15,
-} as const;
-
-/** Debug mode flag (set to true to visualize layout zones) */
-export const DEBUG_LAYOUT_ZONES = false;
-
-// ============================================================================
-// SHADOW PARAMS (premium outer shadow spec)
-// Based on: 24% opacity, 23.3pt blur, 14.3pt distance, 74° angle
-// ============================================================================
-
-export const SHADOW_PARAMS = {
-  /** Shadow type: outer (standard drop shadow) */
-  type: 'outer' as const,
-  /** Angle in degrees (74° = bottom-right bias) */
-  angle: 74,
-  /** Blur radius in points (23pt ≈ soft premium shadow) */
-  blur: 23,
-  /** Offset/distance in points (14pt) */
-  offset: 14,
-  /** Opacity (0.24 = 24%) */
-  opacity: 0.24,
-} as const;
-
-// ============================================================================
-// HELPER FUNCTIONS
-// ============================================================================
-
-/**
- * Normalize title text to force SINGLE LINE
- * - Trim whitespace
- * - Replace all newlines with space
- * - Collapse multiple spaces into one
- * => Result: title NEVER contains line breaks, NEVER forces 2 lines
- */
-export function normalizeTitleText(text: string): string {
-  return text
-    .trim()
-    .replace(/\n+/g, ' ')      // Replace ALL newlines with space
-    .replace(/\s+/g, ' ')      // Collapse multiple spaces into one
-    .trim();                    // Final trim
-}
-
-/**
- * Calculate the visual height of title text
- * Since title is FORCED to single line after normalization,
- * height = 1 * fontSize * lineHeightMultiple
- */
-export function calculateTitleTextHeight(
-  _titleText: string,
-  fontSize: number,
-  lineHeightMultiple: number = 1.15
-): number {
-  // Title is always 1 line after normalization
-  const lines = 1;
-  
-  // Convert font size from points to inches (1 point = 1/72 inch)
-  const fontSizeInInches = fontSize / 72;
-  const textHeight = lines * fontSizeInInches * lineHeightMultiple;
-  
-  return textHeight;
-}
-
-/**
- * Draw debug borders around layout zones and reference lines (when DEBUG_LAYOUT_ZONES is true)
- */
-export function drawDebugLayoutZones(
-  slide: PptxGenJS.Slide,
-  variant: 'chapter' | 'content',
-  titleTextBottomY?: number,
-  accentLineY?: number
-): void {
-  if (!DEBUG_LAYOUT_ZONES) return;
-  
-  const layoutZones = variant === 'chapter' ? LAYOUT_ZONES.chapter : LAYOUT_ZONES.content;
-  const debugColorRed = 'FF0000';    // Red for boxes
-  const debugColorGreen = '00FF00'; // Green for titleTextBottomY
-  const debugColorBlue = '0000FF';  // Blue for accentLineY
-  
-  // Draw title box border (RED)
-  slide.addShape('rect', {
-    x: layoutZones.titleBox.x,
-    y: layoutZones.titleBox.y,
-    w: layoutZones.titleBox.w,
-    h: layoutZones.titleBox.h,
-    fill: { color: 'FFFFFF', transparency: 80 },
-    line: { color: debugColorRed, width: 0.5 },
-  });
-  
-  // Draw subtitle box border (RED)
-  slide.addShape('rect', {
-    x: layoutZones.subtitleBox.x,
-    y: layoutZones.subtitleBox.y,
-    w: layoutZones.subtitleBox.w,
-    h: layoutZones.subtitleBox.h,
-    fill: { color: 'FFFFFF', transparency: 80 },
-    line: { color: debugColorRed, width: 0.5 },
-  });
-  
-  // Draw reference line at titleTextBottomY (GREEN) - proves position is TEXT-based
-  if (titleTextBottomY !== undefined) {
-    slide.addShape('line', {
-      x: layoutZones.titleBox.x,
-      y: titleTextBottomY,
-      w: layoutZones.titleBox.w,
-      h: 0,
-      line: { color: debugColorGreen, width: 0.25, dashType: 'dash' },
-    });
-  }
-  
-  // Draw reference line at accentLineY (BLUE) - shows where bar will be
-  if (accentLineY !== undefined) {
-    slide.addShape('line', {
-      x: layoutZones.titleBox.x,
-      y: accentLineY,
-      w: layoutZones.titleBox.w,
-      h: 0,
-      line: { color: debugColorBlue, width: 0.25, dashType: 'dash' },
-    });
-  }
-}
-export function roleColor(theme: PptxThemeRoles, role: keyof PptxThemeRoles): string {
-  const color = theme[role];
-  if (typeof color === 'string') {
-    return color.replace('#', '');
-  }
-  throw new Error(`Invalid color role: ${String(role)}`);
-}
-
-/**
- * Add solid background to slide
- */
-export function addBackground(slide: PptxGenJS.Slide, color: string): void {
-  slide.background = { color: color.replace('#', '') };
-}
-
-/**
- * Add accent line (horizontal) - POSITIONED BETWEEN TITLE AND SUBTITLE
- * 
- * IMPORTANT: The line is now positioned at the MIDPOINT between title and subtitle zones,
- * not at the bottom of the title box. This creates proper visual separation.
- * 
- * @deprecated Use addHeader() instead for text-based positioning
- */
-export function addAccentLine(
-  slide: PptxGenJS.Slide,
-  theme: PptxThemeRoles,
-  variant: 'cover' | 'chapter' | 'content'
-): void {
-  let coords: { x: number; y: number; w: number };
-  let lineWidth: number;
-  
-  if (variant === 'cover') {
-    // Cover slide uses fixed coordinates (no subtitle zone)
-    coords = COORDS_COVER.dividerLine;
-    lineWidth = 0.75;
-  } else {
-    // Chapter and Content: calculate Y position at midpoint between title and subtitle
-    const layoutZones = variant === 'chapter' ? LAYOUT_ZONES.chapter : LAYOUT_ZONES.content;
-    const baseCoords = variant === 'chapter' ? COORDS_CHAPTER.accentLine : COORDS_CONTENT.accentLine;
-    
-    // Calculate midpoint between title bottom and subtitle top
-    const titleBottomY = layoutZones.titleBox.y + layoutZones.titleBox.h;
-    const subtitleTopY = layoutZones.subtitleBox.y;
-    const accentLineY = titleBottomY + (subtitleTopY - titleBottomY) * 0.5;
-    
-    coords = {
-      x: baseCoords.x,
-      y: accentLineY,
-      w: baseCoords.w
-    };
-    lineWidth = 1.5;
-  }
-  
-  slide.addShape('line', {
-    x: coords.x,
-    y: coords.y,
-    w: coords.w,
-    h: 0,
-    line: {
-      color: roleColor(theme, 'accent'),
-      width: lineWidth,
-    },
-  });
-}
-
-/**
- * Add complete header (title + accent line + subtitle) with TEXT-BASED positioning
- * 
- * IMPORTANT: This function positions the accent line based on the ACTUAL height of the title text,
- * not the title box height. This prevents the line from being "stuck" when the title box is too large
- * or contains empty lines.
- * 
- * @param slide - PptxGenJS slide
- * @param titleText - Title text (will be normalized)
- * @param subtitleText - Subtitle text
- * @param theme - Theme roles for colors
- * @param variant - 'chapter' or 'content' layout
- * @param titleFontSize - Font size for title (default: TYPO.sizes.h1)
- * @param subtitleFontSize - Font size for subtitle (default: TYPO.sizes.h2)
- */
-export function addHeader(
-  slide: PptxGenJS.Slide,
-  titleText: string,
-  subtitleText: string,
-  theme: PptxThemeRoles,
-  variant: 'chapter' | 'content',
-  titleFontSize: number = TYPO.sizes.h1,
-  subtitleFontSize: number = TYPO.sizes.h2
-): void {
-  const layoutZones = variant === 'chapter' ? LAYOUT_ZONES.chapter : LAYOUT_ZONES.content;
-  const baseCoords = variant === 'chapter' ? COORDS_CHAPTER : COORDS_CONTENT;
-  
-  // ========== B1: Normalize title to SINGLE LINE ==========
-  const normalizedTitleText = normalizeTitleText(titleText);
-  
-  // ========== B2: Calculate Y of accent line CENTERED ==========
-  // Title text height (always 1 line after normalization)
-  const titleTextHeight = calculateTitleTextHeight(normalizedTitleText, titleFontSize);
-  
-  // Bottom of title TEXT (not box)
-  const titleTextBottomY = layoutZones.titleBox.y + titleTextHeight;
-  
-  // Top of subtitle (fixed position from layout)
-  const subtitleTopY = layoutZones.subtitleBox.y;
-  
-  // Accent line Y = CENTERED between title text bottom and subtitle top
-  const accentLineY = titleTextBottomY + (subtitleTopY - titleTextBottomY) * 0.5;
-  
-  // ========== B3: Ensure subtitle stays below accent line ==========
-  const minGapAfterLine = 0.08; // minimum gap after accent line
-  const subtitleY = Math.max(subtitleTopY, accentLineY + minGapAfterLine);
-  
-  // ========== C1: Debug mode logging and visual markers ==========
-  if (DEBUG_LAYOUT_ZONES) {
-    // eslint-disable-next-line no-console
-    console.log('[DEBUG HEADER]', {
-      slideType: variant,
-      titleOriginal: titleText,
-      titleNormalized: normalizedTitleText,
-      titleBox: layoutZones.titleBox,
-      subtitleBox: layoutZones.subtitleBox,
-      titleTextBottomY: titleTextBottomY.toFixed(3),
-      subtitleTopY: subtitleTopY.toFixed(3),
-      accentLineY: accentLineY.toFixed(3),
-      subtitleY: subtitleY.toFixed(3),
-      gap_title_to_line: (accentLineY - titleTextBottomY).toFixed(3),
-      gap_line_to_subtitle: (subtitleY - accentLineY).toFixed(3),
-    });
-  }
-  
-  // Draw debug zones with reference lines
-  drawDebugLayoutZones(slide, variant, titleTextBottomY, accentLineY);
-  
-  // ========== Render title ==========
-  addTextBox(slide, normalizedTitleText, layoutZones.titleBox, {
-    fontSize: titleFontSize,
-    color: theme.textMain,
-    bold: true,
-    align: 'left',
-    valign: 'top',
-    isUpperCase: true,
-  });
-  
-  // ========== Render accent line (CENTERED) ==========
-  slide.addShape('line', {
-    x: baseCoords.accentLine.x,
-    y: accentLineY,
-    w: baseCoords.accentLine.w,
-    h: 0,
-    line: {
-      color: roleColor(theme, 'accent'),
-      width: 1.5,
-    },
-  });
-  
-  // ========== Render subtitle (below accent line) ==========
-  const subtitleRect = {
-    x: layoutZones.subtitleBox.x,
-    y: subtitleY,
-    w: layoutZones.subtitleBox.w,
-    h: layoutZones.subtitleBox.h,
-  };
-  
-  addTextBox(slide, subtitleText, subtitleRect, {
-    fontSize: subtitleFontSize,
-    color: theme.textMain,
-    bold: true,
-    align: 'left',
-    valign: 'top',
-  });
-}
-
-/**
- * Add corner marks (vertical lines)
- */
-export function addCornerMarks(
-  slide: PptxGenJS.Slide,
-  theme: PptxThemeRoles,
-  placementVariant: 'coverBottom' | 'endDiagonal'
-): void {
-  const accentColor = roleColor(theme, 'accent');
-  const lineWidth = 0.75;
-  
-  if (placementVariant === 'coverBottom') {
-    // Bottom left corner mark (2 vertical lines)
-    const left = COORDS_COVER.cornerMarkLeft;
-    slide.addShape('line', {
-      x: left.x,
-      y: left.y,
-      w: 0,
-      h: left.h * 0.6,
-      line: { color: accentColor, width: lineWidth },
-    });
-    slide.addShape('line', {
-      x: left.x + 0.15,
-      y: left.y + 0.1,
-      w: 0,
-      h: left.h * 0.5,
-      line: { color: accentColor, width: lineWidth },
-    });
-    
-    // Bottom right corner mark (2 vertical lines)
-    const right = COORDS_COVER.cornerMarkRight;
-    slide.addShape('line', {
-      x: right.x + right.w - 0.15,
-      y: right.y,
-      w: 0,
-      h: right.h * 0.6,
-      line: { color: accentColor, width: lineWidth },
-    });
-    slide.addShape('line', {
-      x: right.x + right.w,
-      y: right.y + 0.1,
-      w: 0,
-      h: right.h * 0.5,
-      line: { color: accentColor, width: lineWidth },
-    });
-  } else if (placementVariant === 'endDiagonal') {
-    // Compute symmetric positions from CORNER_MARKS constants
-    const cm = CORNER_MARKS;
-    const cmLineWidth = cm.lineWidth;
-    
-    // SYMMETRIC POSITIONING:
-    // Top right corner: anchor at top-right, lines go DOWN
-    // x = slideWidth - marginX (right edge of lines)
-    // y = marginY (top of lines)
-    const topRightX = SLIDE_SIZE.width - cm.marginX;
-    const topRightY = cm.marginY;
-    
-    // Bottom left corner: anchor at bottom-left, lines go DOWN from anchor
-    // x = marginX (left edge of lines)
-    // y = slideHeight - marginY - primaryHeight (so bottom of line touches marginY from bottom)
-    const bottomLeftX = cm.marginX;
-    const bottomLeftY = SLIDE_SIZE.height - cm.marginY - cm.primaryHeight;
-    
-    // Top right corner marks (2 vertical lines) - positioned from right edge
-    slide.addShape('line', {
-      x: topRightX - cm.lineSpacing,
-      y: topRightY,
-      w: 0,
-      h: cm.primaryHeight,
-      line: { color: accentColor, width: cmLineWidth },
-    });
-    slide.addShape('line', {
-      x: topRightX,
-      y: topRightY + cm.secondaryOffset,
-      w: 0,
-      h: cm.secondaryHeight,
-      line: { color: accentColor, width: cmLineWidth },
-    });
-    
-    // Bottom left corner marks (2 vertical lines) - positioned from left edge
-    slide.addShape('line', {
-      x: bottomLeftX,
-      y: bottomLeftY,
-      w: 0,
-      h: cm.primaryHeight,
-      line: { color: accentColor, width: cmLineWidth },
-    });
-    slide.addShape('line', {
-      x: bottomLeftX + cm.lineSpacing,
-      y: bottomLeftY + cm.secondaryOffset,
-      w: 0,
-      h: cm.secondaryHeight,
-      line: { color: accentColor, width: cmLineWidth },
-    });
-  }
-}
-
-/**
- * Format date for footer (fr-FR locale)
- */
-function formatDateFr(date: Date): string {
-  const day = date.getDate().toString().padStart(2, '0');
-  const month = (date.getMonth() + 1).toString().padStart(2, '0');
-  const year = date.getFullYear();
-  return `${day}/${month}/${year}`;
-}
-
-/**
- * Add footer component
- */
-export function addFooter(
-  slide: PptxGenJS.Slide,
-  ctx: {
-    generatedAt: Date;
-    footerDisclaimer?: string;
-    showSlideNumbers?: boolean;
-    theme: PptxThemeRoles;
-  },
-  slideIndex: number,
-  variant: 'onLight' | 'accent' | 'onMain'
-): void {
-  let textColor: string;
-  if (variant === 'onLight') {
-    textColor = roleColor(ctx.theme, 'footerOnLight');
-  } else if (variant === 'onMain') {
-    // Use adaptive text color for colored backgrounds
-    textColor = ctx.theme.textOnMain.replace('#', '');
-  } else {
-    textColor = roleColor(ctx.theme, 'footerAccent');
-  }
-  
-  // Use explicit check to allow empty string to suppress disclaimer
-  const disclaimer = ctx.footerDisclaimer !== undefined 
-    ? ctx.footerDisclaimer 
-    : "Document non contractuel établi en fonction des dispositions fiscales ou sociales en vigueur à la date des présentes";
-  
-  // Date
-  addTextFr(slide, formatDateFr(ctx.generatedAt), {
-    x: COORDS_FOOTER.date.x,
-    y: COORDS_FOOTER.date.y,
-    w: COORDS_FOOTER.date.w,
-    h: COORDS_FOOTER.date.h,
-    fontSize: TYPO.sizes.footer,
-    fontFace: TYPO.fontFace,
-    color: textColor,
-    align: 'left',
-    valign: 'middle',
-  });
-  
-  // Disclaimer
-  addTextFr(slide, disclaimer, {
-    x: COORDS_FOOTER.disclaimer.x,
-    y: COORDS_FOOTER.disclaimer.y,
-    w: COORDS_FOOTER.disclaimer.w,
-    h: COORDS_FOOTER.disclaimer.h,
-    fontSize: TYPO.sizes.footer,
-    fontFace: TYPO.fontFace,
-    color: textColor,
-    align: 'center',
-    valign: 'middle',
-  });
-  
-  // Slide number
-  const slideNumText = ctx.showSlideNumbers !== false ? `${slideIndex}` : 'N°';
-  addTextFr(slide, slideNumText, {
-    x: COORDS_FOOTER.slideNum.x,
-    y: COORDS_FOOTER.slideNum.y,
-    w: COORDS_FOOTER.slideNum.w,
-    h: COORDS_FOOTER.slideNum.h,
-    fontSize: TYPO.sizes.footer,
-    fontFace: TYPO.fontFace,
-    color: textColor,
-    align: 'right',
-    valign: 'middle',
-  });
-}
-
-/**
- * Add card panel with premium shadow effect
- * 
- * Uses native PPTXGenJS shadow property for a single clean shadow.
- * NO multi-layer simulation - one shape only.
- * 
- * Panel specs:
- * - Fill: white (#FFFFFF)
- * - Border: color 8 (panelBorder role), 0.75pt width
- * - Shadow: outer, 24% opacity, 23pt blur, 14pt offset, 74° angle
- * 
- * @param slide - PptxGenJS slide
- * @param rect - Panel bounding box
- * @param theme - Theme roles for colors
- * @param panelRadius - Corner radius for roundRect
- */
-export function addCardPanelWithShadow(
-  slide: PptxGenJS.Slide,
-  rect: { x: number; y: number; w: number; h: number },
-  theme: PptxThemeRoles,
-  panelRadius: number = RADIUS.panel
-): void {
-  const shadowColor = roleColor(theme, 'shadowBase');
-  const borderColor = roleColor(theme, 'panelBorder');
-  
-  // Single panel with native PPTXGenJS shadow
-  // NO multi-layer simulation - clean single shape
-  slide.addShape('roundRect', {
-    x: rect.x,
-    y: rect.y,
-    w: rect.w,
-    h: rect.h,
-    fill: { color: 'FFFFFF' },
-    line: { color: borderColor, width: 0.75 },
-    rectRadius: panelRadius,
-    shadow: {
-      type: SHADOW_PARAMS.type,
-      angle: SHADOW_PARAMS.angle,
-      blur: SHADOW_PARAMS.blur,
-      offset: SHADOW_PARAMS.offset,
-      opacity: SHADOW_PARAMS.opacity,
-      color: shadowColor,
-    },
-  });
-}
-
-/**
- * Legacy function - kept for backward compatibility
- * @deprecated Use addCardPanelWithShadow instead
- */
-export function addPanelWithShadow(
-  slide: PptxGenJS.Slide,
-  rect: { x: number; y: number; w: number; h: number },
-  theme: PptxThemeRoles
-): void {
-  addCardPanelWithShadow(slide, rect, theme);
-}
-
-/**
- * Add chapter image from public assets with bleed
- * 
- * IMPORTANT: Images are PRE-PROCESSED with rounded corners (in /public/pptx/chapters/)
- * Do NOT apply rounding here - it creates an oval/ellipse effect which is incorrect.
- * 
- * BLEED: Image extends slightly under the panel border to eliminate
- * the "hole" caused by anti-aliasing at rounded corners.
- */
-export function addChapterImage(
-  slide: PptxGenJS.Slide,
-  imageDataUri: string,
-  rect: { x: number; y: number; w: number; h: number },
-  applyBleed: boolean = true
-): void {
-  // Apply bleed to eliminate anti-aliasing gaps at corners
-  const bleed = applyBleed ? BLEED.image : 0;
-  
-  slide.addImage({
-    data: imageDataUri,
-    x: rect.x - bleed,
-    y: rect.y - bleed,
-    w: rect.w + bleed, // Only extend right edge (left is panel edge)
-    h: rect.h + 2 * bleed,
-    // NO rounding option - image is pre-processed with rounded corners
-  });
-}
-
-/**
- * Validate that a rect is within slide bounds (NO OVERFLOW)
- * @throws Error if rect exceeds slide canvas
- */
-export function validateNoOverflow(
-  rect: { x: number; y: number; w: number; h: number },
-  context: string = 'element'
-): void {
-  const rightEdge = rect.x + rect.w;
-  const bottomEdge = rect.y + rect.h;
-  
-  if (rect.x < 0 || rect.y < 0) {
-    if (import.meta.env.DEV) {
-      console.warn(`[Layout Contract] ${context}: Negative position detected (x:${rect.x}, y:${rect.y})`);
-    }
-  }
-  if (rightEdge > SLIDE_SIZE.width) {
-    if (import.meta.env.DEV) {
-      console.warn(`[Layout Contract] ${context}: Right edge overflow (${rightEdge} > ${SLIDE_SIZE.width})`);
-    }
-  }
-  if (bottomEdge > SLIDE_SIZE.height) {
-    if (import.meta.env.DEV) {
-      console.warn(`[Layout Contract] ${context}: Bottom edge overflow (${bottomEdge} > ${SLIDE_SIZE.height})`);
-    }
-  }
-}
-
-/**
- * Clamp a rect to stay within slide bounds
- * Used to enforce NO OVERFLOW rule
- */
-export function clampToSlide(
-  rect: { x: number; y: number; w: number; h: number }
-): { x: number; y: number; w: number; h: number } {
-  const x = Math.max(0, rect.x);
-  const y = Math.max(0, rect.y);
-  const w = Math.min(rect.w, SLIDE_SIZE.width - x);
-  const h = Math.min(rect.h, SLIDE_SIZE.height - y);
-  return { x, y, w, h };
-}
-
-/**
- * Add text with enforced French proofing and default Arial font
- * Use this helper instead of slide.addText directly.
- */
-export function addTextFr(
-  slide: PptxGenJS.Slide,
-  text: string | PptxGenJS.TextProps[],
-  options: PptxGenJS.TextPropsOptions & Record<string, unknown>
-): void {
-  const normalizedText = Array.isArray(text)
-    ? text.map((run) => ({
-        ...run,
-        options: {
-          ...run.options,
-          lang: 'fr-FR',
-          fontFace: run.options?.fontFace || TYPO.fontFace,
-        },
-      }))
-    : text;
-  slide.addText(normalizedText, {
-    ...options,
-    fontFace: options.fontFace || TYPO.fontFace,
-    lang: 'fr-FR',
-  });
-}
-
-/**
- * Add text box with consistent defaults
- * Enforces NO OVERFLOW by clamping to slide bounds
- */
-export function addTextBox(
-  slide: PptxGenJS.Slide,
-  text: string,
-  rect: { x: number; y: number; w: number; h: number },
-  style: {
-    fontSize?: number;
-    fontFace?: string;
-    color: string;
-    bold?: boolean;
-    align?: 'left' | 'center' | 'right';
-    valign?: 'top' | 'middle' | 'bottom';
-    isUpperCase?: boolean;
-    lineSpacing?: number; // Line spacing multiplier (e.g., 1.15 for 115%)
-    wrap?: boolean; // Enable text wrapping (default: true)
-  }
-): void {
-  const displayText = style.isUpperCase ? text.toUpperCase() : text;
-  
-  // Validate and clamp to slide bounds (NO OVERFLOW)
-  validateNoOverflow(rect, 'TextBox');
-  const safeRect = clampToSlide(rect);
-  
-  // Build text options with French proofing language
-  const textOptions: PptxGenJS.TextPropsOptions = {
-    x: safeRect.x,
-    y: safeRect.y,
-    w: safeRect.w,
-    h: safeRect.h,
-    fontSize: style.fontSize || TYPO.sizes.body,
-    fontFace: style.fontFace || TYPO.fontFace,
-    color: style.color.replace('#', ''),
-    bold: style.bold || false,
-    align: style.align || 'left',
-    valign: style.valign || 'top',
-    wrap: style.wrap !== false, // Default to true for text fitting
-    lang: 'fr-FR', // French proofing language
-  };
-  
-  // Add line spacing if specified (PptxGenJS uses lineSpacingMultiple)
-  if (style.lineSpacing) {
-    textOptions.lineSpacingMultiple = style.lineSpacing;
-  }
-  
-  slide.addText(displayText, textOptions);
-}
-
-// ============================================================================
-// CHAPTER IMAGE POOLS
-// ============================================================================
-
-/**
- * Pool d'images chapitres par simulateur.
- * Chaque simulateur a ses propres images pour éviter les collisions visuelles.
- * Images disponibles : public/pptx/chapters/ch-01.png .. ch-09.png (range strict 1–9).
  *
- * Matrice de diversité :
- *   ch-01 → IR
- *   ch-02 → Succession, Placement
- *   ch-03 → IR
- *   ch-05 → Credit
- *   ch-06 → Succession
- *   ch-07 → Credit
- *   ch-09 → Placement
+ * Ce fichier est un barrel re-export — les symboles sont définis dans les
+ * modules du dossier `designSystem/`.
  */
-export const CHAPTER_IMAGE_POOLS: Record<string, readonly number[]> = {
-  ir:         [1, 3],  // source de vérité — inchangé
-  credit:     [5, 7],  // diversifié vs IR (était [1, 3])
-  per:        [4, 8],  // potentiel PER
-  succession: [2, 6],  // diversifié (était [1, 3])
-  placement:  [9, 2],  // futur simulateur
-} as const;
 
-/**
- * Retourne l'index d'image de chapitre pour un simulateur et un ordinal donné.
- * Utiliser à la place d'un index hardcodé dans les deck builders.
- *
- * @param simId - Identifiant du simulateur ('ir', 'credit', 'succession', 'per', 'placement')
- * @param chapterOrdinal - Ordinal 0-based dans le deck (0 = premier chapitre, 1 = second, etc.)
- * @returns Index d'image valide (1–9)
- */
-export function pickChapterImage(simId: string, chapterOrdinal: number): number {
-  const pool = CHAPTER_IMAGE_POOLS[simId];
-  if (!pool || pool.length === 0) return 1;
-  return pool[chapterOrdinal % pool.length];
-}
-
-export default {
-  SLIDE_SIZE,
-  TYPO,
-  RADIUS,
-  BLEED,
-  CORNER_MARKS,
-  LAYOUT_ZONES,
-  MIN_FONT_SIZES,
-  TEXT_GAPS,
-  COORDS_COVER,
-  COORDS_CHAPTER,
-  COORDS_CONTENT,
-  COORDS_END,
-  COORDS_FOOTER,
-  LINE_STYLES,
-  SHADOW_PARAMS,
-  LOGO_PLACEMENT,
-  DEFAULT_LOGO_PLACEMENT,
-  roleColor,
-  addBackground,
-  addAccentLine,
-  addHeader,
-  addCornerMarks,
-  addFooter,
-  addCardPanelWithShadow,
-  addPanelWithShadow,
-  addChapterImage,
-  addTextFr,
-  addTextBox,
-  normalizeTitleText,
-  calculateTitleTextHeight,
-  drawDebugLayoutZones,
-  validateNoOverflow,
-  clampToSlide,
-  calculateLogoPosition,
-  CHAPTER_IMAGE_POOLS,
-  pickChapterImage,
-};
+export * from './layout';
+export * from './typography';
+export * from './coords';
+export * from './logo';
+export * from './chapters';
+export * from './colorsShadows';
+export * from './helpers';
+export * from './components';
